ApplicationLogicLayer/Features/Foo/AddService.cs
using AutoMapper;
using Core.Extensions;
using DataAccessLayer;
using DataAccessLayer.Models;
using EntityFramework.DbContextScope.Interfaces;
using FluentValidation;
using MediatR;
using System.Threading;
using System.Threading.Tasks;

namespace ApplicationLogicLayer.Features.Foo
{
    public class AddService
    {
        public class Command : IRequest<Unit>
        {
            public string Name { get; set; }
            public string Description { get; set; }
        }

        public class MappingProfile : Profile
        {
            public MappingProfile()
            {
                CreateMap<Command, Bar>().SetId();
            }
        }

        public class CommandValidator : AbstractValidator<Command>
        {
            public CommandValidator()
            {
                RuleFor(r => r.Name).NotNull().NotEmpty().WithGlobalMessage("Name is required");
                RuleFor(r => r.Description).NotNull().NotEmpty().WithGlobalMessage("Description is required");
            }
        }

        public class CommandHandler : IRequestHandler<Command>
        {
            private readonly IMapper _mapper;
            private readonly IDbContextScopeFactory _scopeFactory;

            public CommandHandler(IMapper mapper, IDbContextScopeFactory scopeFactory)
            {
                _mapper = mapper;
                _scopeFactory = scopeFactory;
            }

            public Task<Unit> Handle(Command request, CancellationToken cancellationToken)
            {
                using (var scope = _scopeFactory.Create())
                {
                    var context = scope.DbContexts.Get<AppDbContext>();

                    var itemToAdd = _mapper.Map<Bar>(request);

                    context.Set<Bar>().Add(itemToAdd);

                    scope.SaveChanges();
                }

                return Task.FromResult(Unit.Value);
            }
        }
    }
}
